Delete duplicated functions. Fix tests.

+++ model/rbac_test.go
@@ -113,7 +113,7 @@ func TestRBAC(t *testing.T) {
 		if err != nil {
 			t.Error(err)
 		} else {
-			assert.Equal(t, uint8(1), permissions[0].Type)
+			assert.Equal(t, uint(1), permissions[0].Type)
 		}
 
 		if err := tRepo.mock.ExpectationsWereMet(); err != nil {
